requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pypdf>=3.17.0
python-docx>=1.0.0

# Utilities
python-magic>=0.4.27; sys_platform != 'win32'
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Progress and UI
tqdm>=4.66.0
